Renamed ChampionButton.js to CoursesButton.js to match the new button name
